+++ src/const/index.ts
@@ -1,0 +1,3 @@
+import regMap from './reg-map';
+
+export default { reg: regMap };