+++ src/const/reg-map.ts
@@ -1,0 +1,6 @@
+// hex 正则
+const Hex = /^#([a-fA-F0-9]{3}|[a-fA-F0-9]{6})$/;
+
+export default {
+  Hex,
+};