+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './type-checks';
 export * from './common';
+export * from './numbers';