+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOGS
 
+## 0.0.5
+* Add Function method `debounce` `throttle`
+* Add `hex2Rgba` method
+
 ## 0.0.4
 * Add `sliceLoad` method
 * Add `equal` method
@@ -9,4 +13,4 @@
 ## 0.0.3
 * Add `type-checks` module
 * Add `copyOf` `copyDeepOf` `copyDeep` method
-* Add `slice`  method
+* Add `slice` method